iter1/fix/init
