+++ src/components/features/index.ts
@@ -1,0 +1,7 @@
+import TextSize from "./TextSize";
+import OpacityText from "./OpacityText";
+import Ratio from "./Ratio";
+import FontText from "./FontText";
+import ColorText from "./ColorText";
+
+export { TextSize, OpacityText, Ratio, FontText, ColorText}
