+++ src/config/env.ts
@@ -1,0 +1,4 @@
+const SERVER_PORT = 'http://10.0.2.2:9090/api';
+
+
+export {SERVER_PORT}
