+++ src/utils/data.ts
@@ -1,0 +1,64 @@
+const DataFont: string[] = ['Audiowide-Regular',
+    'DMSeriText-Italic',
+    'DMSeriText-Regular',
+    'PlayfairDisplay-Black',
+    'PlayfairDisplay-BlackItalic',
+    'PlayfairDisplay-Bold',
+    'PlayfairDisplay-BoldItalic',
+    'PlayfairDisplay-ExtraBold',
+    'PlayfairDisplay-ExtraBoldItalic',
+    'PlayfairDisplay-Italic',
+    'PlayfairDisplay-Medium',
+    'PlayfairDisplay-MediumItalic',
+    'PlayfairDisplay-Regular',
+    'PlayfairDisplay-SemiBold',
+    'PlayfairDisplay-SemiBoldItalic',
+    'PlaywriteAUSA-ExtraLight',
+    'PlaywriteAUSA-Light',
+    'PlaywriteAUSA-Regular',
+    'PlaywriteAUSA-Thin',
+    'PlaywriteIN-ExtraLight',
+    'PlaywriteIN-Light',
+    'PlaywriteIN-Regular',
+    'PlaywriteIN-Thin',
+    'PlaywriteVN-ExtraLight',
+    'PlaywriteVN-Light',
+    'PlaywriteVN-Regular',
+    'PlaywriteVN-Thin',
+]
+
+const DataColor: string[] = [
+    "#FF0000",
+    "#FF4500",
+    "#FF6347",
+    "#FF7F50",
+    "#FFA07A",
+    "#FFD700",
+    "#FFFF00",
+    "#FFFFE0",
+    "#00FF00",
+    "#32CD32",
+    "#98FB98",
+    "#00FFFF",
+    "#4682B4",
+    "#1E90FF",
+    "#87CEFA",
+    "#0000FF",
+    "#8A2BE2",
+    "#9400D3",
+    "#BA55D3",
+    "#DDA0DD",
+    "#FF1493",
+    "#FF69B4",
+    "#FFB6C1",
+    "#800000",
+    "#A52A2A",
+    "#D2691E",
+    "#F4A460",
+    "#FFF5EE",
+    "#708090",
+    "#D3D3D3"
+];
+
+
+export { DataFont, DataColor }
